Update README and make predictions with floating point output

+++ lib/helper/isolate_inference.dart
@@ -77,14 +77,14 @@ class IsolateInference {
       // Set tensor input [1, 224, 224, 3]
       final input = [imageMatrix];
       // Set tensor output [1, 1001]
-      final output = [List<int>.filled(isolateModel.outputShape[1], 0)];
+      final output = [List<double>.filled(isolateModel.outputShape[1], 0)];
       // // Run inference
       Interpreter interpreter =
           Interpreter.fromAddress(isolateModel.interpreterAddress);
       interpreter.run(input, output);
       // Get first output tensor
       final result = output.first;
-      int maxScore = result.reduce((a, b) => a + b);
+      double maxScore = result.reduce((a, b) => a + b);
       // Set classification map {label: points}
       var classification = <String, double>{};
       for (var i = 0; i < result.length; i++) {
